Modularized route handling into separate urlRoutes.js file for cleaner code structure.
